Made those refactor changes to be more explicit to the system calls as well

+++ src/main/kotlin/helpers/readRegisterString.kt
@@ -10,7 +10,7 @@ fun readRegisterString(register: SuperRegisterType): String {
     while (true) {
         val byte = internalMemory.read(MemoryAddress(fullRegisterRead(register)!! + index))
         if (byte.value == 0L) break
-        string += byte.value!!.toChar()
+        string += byte.value!!.toInt().toChar()
         index++
     }
     return string
